Move device drivers from lib
