+++ drivers/light/TuyaLightConstants.js
