+++ drivers/sensor_contact/driver.js
@@ -1,3 +1,31 @@
 'use strict';
 
-module.exports = require('../../lib/TuyaOAuth2DriverSensorContact');
+const TuyaOAuth2DriverSensor = require('../../lib/TuyaOAuth2DriverSensor');
+const TuyaOAuth2Constants = require('../../lib/TuyaOAuth2Constants');
+
+/**
+ * @extends TuyaOAuth2DriverSensor
+ * @hideconstructor
+ */
+class TuyaOAuth2DriverSensorContact extends TuyaOAuth2DriverSensor {
+
+  static TUYA_DEVICE_CATEGORIES = [
+    TuyaOAuth2Constants.DEVICE_CATEGORIES.SECURITY_VIDEO_SURV.CONTACT_SENSOR,
+  ];
+
+  onTuyaPairListDeviceProperties(device) {
+    const props = super.onTuyaPairListDeviceProperties(device);
+
+    // alarm_contact
+    const hasDoorContactState = device.status.some(({ code }) => code === 'doorcontact_state');
+    if (hasDoorContactState) {
+      props.store.tuya_capabilities.push('doorcontact_state');
+      props.capabilities.push('alarm_contact');
+    }
+
+    return props;
+  }
+
+}
+
+module.exports = TuyaOAuth2DriverSensorContact;
