+++ drivers/socket/TuyaSocketConstants.js
