--- lib/TuyaOAuth2DriverSensorSmoke.js
@@ -1,31 +1,0 @@
-'use strict';
-
-const TuyaOAuth2DriverSensor = require('./TuyaOAuth2DriverSensor');
-const TuyaOAuth2Constants = require('./TuyaOAuth2Constants');
-
-/**
- * @extends TuyaOAuth2DriverSensor
- * @hideconstructor
- */
-class TuyaOAuth2DriverSensorSmoke extends TuyaOAuth2DriverSensor {
-
-  static TUYA_DEVICE_CATEGORIES = [
-    TuyaOAuth2Constants.DEVICE_CATEGORIES.SECURITY_VIDEO_SURV.SMOKE_ALARM,
-  ];
-
-  onTuyaPairListDeviceProperties(device) {
-    const props = super.onTuyaPairListDeviceProperties(device);
-
-    // alarm_smoke
-    const hasSmokeSensorStatus = device.status.some(({ code }) => code === 'smoke_sensor_status');
-    if (hasSmokeSensorStatus) {
-      props.store.tuya_capabilities.push('smoke_sensor_status');
-      props.capabilities.push('alarm_smoke');
-    }
-
-    return props;
-  }
-
-}
-
-module.exports = TuyaOAuth2DriverSensorSmoke;
